Fix bug allowing climbing glass walls
Previous attempt to stop wall climbing was not effective on all block types. It allowed climbing against glass walls.
Now the block under check is split in two
- For the block under the player's 'reported' position only a block under is required.
- For the blocks adjacent to that position (where the player may be if crouching at an edge) there must be a block with empty space above it.
No longer bothers to attempt to allow part blocks for the empty space. This was the mistake that allowed glass since it was a bad check for a part-block.

Also, checking for a block under the player is now done only for a player boosting themselves. This ensures that players can't camp in a place where boosting won't work because it would have otherwise been trying to prevent wall climbing with the block-under check.