Fix use of defaults from included resource files
Getting values for known settings was providing a default in the call but this overrode the default provided by calling setDefaults() with the included resource files. This happens even if the value in the call is null. The value from setDefaults() is only used when the method with no def parameter is called.